Move memory id to value AIR
