+++ stwo_cairo_prover/crates/prover/src/components/range_check_builtin/component.rs
@@ -9,7 +9,7 @@ use stwo_prover::core::fields::secure_column::SECURE_EXTENSION_DEGREE;
 use stwo_prover::core::lookups::utils::Fraction;
 use stwo_prover::core::pcs::TreeVec;
 
-use crate::components::memory::{MemoryLookupElements, N_ADDRESS_FELTS, N_BITS_PER_FELT};
+use crate::components::memory::id_to_f252::{MemoryLookupElements, N_ADDRESS_FELTS, N_BITS_PER_FELT};
 use crate::components::LOOKUP_INTERACTION_PHASE;
 use crate::input::SegmentAddrs;
