+++ stwo_cairo_prover/crates/prover/src/components/ret_opcode/component.rs
@@ -9,8 +9,8 @@ use stwo_prover::core::fields::secure_column::SECURE_EXTENSION_DEGREE;
 use stwo_prover::core::lookups::utils::Fraction;
 use stwo_prover::core::pcs::TreeVec;
 
-use crate::components::memory::component::N_M31_IN_FELT252;
-use crate::components::memory::MemoryLookupElements;
+use crate::components::memory::id_to_f252::component::N_M31_IN_FELT252;
+use crate::components::memory::id_to_f252::MemoryLookupElements;
 
 pub const RET_N_TRACE_CELLS: usize = 7;
 pub const RET_INSTRUCTION: [u32; N_M31_IN_FELT252] = [
